backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.0
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.0.0
channels_redis==4.2.1
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
dataclasses-json==0.6.7
Django==3.2.24
django-cors-headers==4.3.1
djangorestframework==3.15.1
dnspython==2.7.0
filetype==1.2.0
frozenlist==1.5.0
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.62.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jsonpatch==1.33
jsonpointer==3.0.0
langsmith==0.3.18
marshmallow==3.26.1
mongoengine==0.27.0
msgpack==1.1.0
multidict==6.2.0
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.16
packaging==23.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.6.2
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
redis==5.0.3
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
service-identity==24.2.0
setuptools==69.1.1
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlparse==0.5.3
tenacity==8.1.0
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.18.3
zope.interface==7.2
zstandard==0.23.0